Reformat x86 code in "Three idioms"; fix atom.xml
Reformat the x86 code in "Three idioms" to make it fit the line limit better.

This change updates the atom.xml feed. At the same time, I hard-coded (for
the moment) &mdash; => &#8212;  The former is not acceptable in XML without
including the appropriate definitions, while the latter works in XML. The
mdash entity caused a validation failure which may have contributed to
Bing not accepting atom.xml as a site map.

In future, probably should fix the original Markdown to use &#8212;
at least if this does get Bing to accept the file.