Use site map with hand-prefixed URLs for Bing
Add an `atom.xml` where every `<link href=` has been hand-prefixed by
`https://tedkirkpatrick.github.io`.  This is a test of whether
Bing will find the pages in the site map when this prefix has been
added.

If this works, then Jekyll's config may have to be changed to ensure
it does the prefixing automatically.